Benchmark a function across different package versions

bench() allows you to run a single function, fn, multiple times in separate R sessions, where each R session has different versions of packages installed. A typical use case is running a before/after benchmark, comparing the CRAN version of a package with a development version of the same package.

For example, bench(fn, pkgs = c("vctrs", "r-lib/vctrs#100")) would run fn in 2 separate R sessions, one with CRAN vctrs installed, and one with the pull request installed.

When using bench(), each call to fn must return a <bench_mark> data frame from the bench package. This is typically generated using either a call to bench::mark() or bench::press().

bench() is similar to run(), but is specifically designed to be useful in conjunction with the bench package when benchmarking across multiple package versions.

bench(
  fn,
  ...,
  pkgs,
  args = list(),
  libpath = .libPaths(),
  args_pak = list(),
  args_callr = list()
)

Arguments

fn

[function]

A function to evaluate. The function is passed along to callr::r(), so it is evaluated in a fresh R session and must be self-contained.

Read the func docs of callr::r() for the full set of restrictions on fn.

fn is converted to a function with rlang::as_function(), so it can be a lambda function.

...

These dots are for future extensions and must be empty.

pkgs

[character]

A character vector of package names or remote package specifications to install and run fn against. Passed along to pak::pkg_install().

args

[list]

An optional list of arguments to pass to the function.

libpath

[character]

The base library path to use.

args_pak

[named list]

A named list of arguments to pass on to pak::pkg_install().

Can't include:

pkg

lib

ask

args_callr

[named list]

A named list of arguments to pass on to callr::r().

Can't include:

func

args

libpath

Value

A <bench_mark>.

Details

The installed packages are placed in temporary directories that act as an extra library path. The temporary directory is prepended to libpath, which is then passed along to callr::r() as the libpath argument there. This ensures that your personal package libraries remain untouched.

Global options

cross.quiet: [TRUE / FALSE]

Should cross specific messages about the process be shown?

Defaults to !rlang::is_interactive() if not set, so messages are only shown when running interactively.

cross.pak_quiet: [TRUE / FALSE]

Should pak installation messages be shown?

Defaults to TRUE if not set.

Examples

if (FALSE) {
# Run a benchmark across 2 different versions of vctrs
bench(pkgs = c("vctrs", "r-lib/vctrs"), ~{
  library(vctrs)
  x <- c(TRUE, FALSE, NA, TRUE)
  bench::mark(vec_detect_missing(x))
})

# You can also use `bench::press()` to generate a grid
bench(pkgs = c("vctrs", "r-lib/vctrs"), ~{
  library(vctrs)
  set.seed(123)
  x <- sample(100)

  bench::press(
    table = c(10, 20, 30),
    {
      table <- seq_len(table)
      bench::mark(vec_match(x, table))
    }
  )
})
}