Evaluate a function across different local package branches

run_branches() is similar to run(), except it allows you to run fn across different local branches corresponding to the same package, rather than different CRAN or GitHub versions of that package.

The default behavior runs the current branch against the main branch.

run_branches(
  fn,
  ...,
  args = list(),
  current = TRUE,
  branches = "main",
  libpath = .libPaths(),
  args_pak = list(),
  args_callr = list()
)

Arguments

fn

[function]

A function to evaluate. The function is passed along to callr::r(), so it is evaluated in a fresh R session and must be self-contained.

Read the func docs of callr::r() for the full set of restrictions on fn.

fn is converted to a function with rlang::as_function(), so it can be a lambda function.

...

These dots are for future extensions and must be empty.

args

[list]

An optional list of arguments to pass to the function.

current

[TRUE / FALSE]

Should the current git branch be included in the vector of branches?

branches

[character]

A character vector of git branch names to check out, install, and run fn against.

It is expected that your working directory is set to the git directory of the package you want to install different branches of. This is typically the case whenever you open an RStudio project for the package in question. Technically, the path is determined by usethis::proj_get().

Your git tree must be completely clean to use branches. If there are any uncommitted changes, an error will be thrown because run() must swap between the branches to install the package, potentially resulting in a loss of information. Note that untracked files are not included in this check - they should never be lost when the branch is changed, but they could affect the results.

After the last branch is installed, the original branch is checked out.

libpath

[character]

The base library path to use.

args_pak

[named list]

A named list of arguments to pass on to pak::pkg_install().

Can't include:

pkg

lib

ask

args_callr

[named list]

A named list of arguments to pass on to callr::r().

Can't include:

func

args

libpath

Value

A data frame with two columns:

branch, a character vector containing branches.

result, a list column containing the result of calling fn for that branch of the package.

Global options

cross.quiet: [TRUE / FALSE]

Should cross specific messages about the process be shown?

Defaults to !rlang::is_interactive() if not set, so messages are only shown when running interactively.

cross.pak_quiet: [TRUE / FALSE]

Should pak installation messages be shown?

Defaults to TRUE if not set.

Examples

if (FALSE) {
# Similar to `run()`, but this runs the function across 2 local branches.
# To run this:
# - The working directory is set to the RStudio project for vctrs
# - There can't be any uncommitted git changes
# - You are currently on a branch, say `fix/performance-bug`
# - You'd like to run that branch against `main`
run_branches(~{
  library(vctrs)
  x <- c(TRUE, FALSE, NA, TRUE)
  bench::mark(vec_detect_missing(x))
})
}