Evaluate an expression across different package versions

run_versions() allows you to run a single expression, expr, multiple times in separate R sessions, where each R sessions has different versions of packages installed. If you're looking to run benchmarks across different versions of packages, you likely want bench_versions() instead.

For example, run_versions(expr, pkgs = c("vctrs", "r-lib/vctrs#100")) would run expr in 2 separate R sessions, one with CRAN vctrs installed, and one with the pull request installed.

run_versions(
  expr,
  ...,
  pkgs,
  libpath = .libPaths(),
  args_pak = list(),
  args_callr = list()
)

Arguments

expr

[expression]

An expression to evaluate. The expression is passed along to callr::r() as the body of a function with zero arguments, so it is evaluated in a fresh R session and must be self-contained.

Read the func docs of callr::r() for the full set of restrictions on expr.

...

These dots are for future extensions and must be empty.

pkgs

[character / data frame of character columns]

A character vector of package names or remote package specifications to install and run expr against. Passed along to pak::pkg_install().

For more complex installations, also accepts a data frame of character columns of package names. Each row in the data frame specifies a package combination to run against.

libpath

[character]

The base library path to use.

args_pak

[named list]

A named list of arguments to pass on to pak::pkg_install().

Can't include:

pkg

lib

ask

args_callr

[named list]

A named list of arguments to pass on to callr::r().

Can't include:

func

args

libpath

Value

A data frame with two columns:

pkg, a character vector containing the packages installed.

result, a list column containing the result of calling expr for that combination of packages.

Details

The installed packages are placed in temporary directories that act as an extra library path. The temporary directory is prepended to libpath, which is then passed along to callr::r() as the libpath argument there. This ensures that your personal package libraries remain untouched.

Global options

cross.quiet: [TRUE / FALSE]

Should cross specific messages about the process be shown?

Defaults to !rlang::is_interactive() if not set, so messages are only shown when running interactively.

cross.pak_quiet: [TRUE / FALSE]

Should pak installation messages be shown?

Defaults to TRUE if not set.

Examples

if (FALSE) {
# Run a benchmark across 2 different versions of vctrs
# (See `bench_versions()` for an even easier way)
run_versions(pkgs = c("vctrs", "r-lib/vctrs"), {
  library(vctrs)
  x <- c(TRUE, FALSE, NA, TRUE)
  bench::mark(vec_detect_missing(x))
})

# Run a benchmark against a combination of development versions
pkgs <- tibble::tribble(
  ~vctrs, ~purrr,
  "vctrs", "purrr",
  "r-lib/vctrs", "purrr",
  "r-lib/vctrs", "tidyverse/purrr"
)

run_versions(pkgs = pkgs, {
  library(purrr)
  x <- list(1, 2)
  bench::mark(map(x, is.double))
})
}